requirements.txt
aniso8601==6.0.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-admin-rangefilter==0.5.0
django-filter==2.2.0
django-graphql-jwt==0.2.1
django-jalali==3.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
graphene==2.1.7
graphene-django==2.4.0
graphene-file-upload==1.2.2
graphql-core==2.2
graphql-relay==2.0.0
idna==2.8
itypes==1.1.0
jdatetime==3.2.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
promise==2.2.1
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
